pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mimic-explorer"
version = "0.1.0"
description = "A locally installable EHR research sandbox for the MIMIC-IV dataset"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "CC-BY-NC-ND-4.0"}
authors = [{name = "Khalid Nawab", email = "khalid.nwb@gmail.com"}]
keywords = ["mimic", "ehr", "fhir", "clinical-data", "research"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "django>=5.0",
    "djangorestframework",
    "django-cors-headers",
    "duckdb>=1.0",
    "fhir.resources>=7.0",
    "orjson",
]

[project.urls]
Homepage = "https://github.com/khalidnawab/mimic-explorer"
Repository = "https://github.com/khalidnawab/mimic-explorer"
Issues = "https://github.com/khalidnawab/mimic-explorer/issues"

[project.scripts]
mimic-explorer = "mimic_explorer.cli:main"

[tool.setuptools.packages.find]
include = ["mimic_explorer*", "core*", "patients*", "clinical*", "fhir*", "research*", "dashboards*"]

[tool.setuptools.package-data]
"*" = ["templates/*.html", "static/**/*"]
